requirements.txt
sqlalchemy==1.4.35
pymysql==1.1.1
pandas==2.2.3
scikit-learn==1.5.2
loguru==0.7.2
python-dotenv==1.0.1
black==24.10.0
boto3==1.35.39
pyarrow==17.0.0